(move) SelectCollateral component
